simple_school_management_system/requirements.txt
fpdf==1.7.2
openpyxl
pillow
pandas
mysql-connector-python
